Changed app.component file for making routing
