+++ app/app.component.ts
@@ -1,96 +1,19 @@
-import {Component, OnInit} from 'angular2/core';
-import {HeroDetailComponent} from './hero-detail.component';
-import {HeroService} from './hero.service';
+import { Component } from 'angular2/core';
+import { HeroService } from './hero.service';
+import { HeroesComponent } from './heroes.component';
 
-// This is metadata for class
 @Component({
-  selector: 'my-app',
-  template: `
-    <h1>{{title}}</h1>
-    <h2>My Heroes</h2>
-    <ul class="heroes">
-      <li *ngFor="#hero of heroes"
-        [class.selected]="hero === selectedHero"
-        (click)="onSelect(hero)">
-        <span class="badge">{{hero.id}}</span> {{hero.name}}
-      </li>
-    </ul>
-    <my-hero-detail [hero]="selectedHero"></my-hero-detail>
-  `,
-  styles: [`
-    .selected {
-      background-color: #CFD8DC !important;
-      color: white;
-    }
-    .heroes {
-      margin: 0 0 2em 0;
-      list-style-type: none;
-      padding: 0;
-      width: 10em;
-    }
-    .heroes li {
-      cursor: pointer;
-      position: relative;
-      left: 0;
-      background-color: #EEE;
-      margin: .5em;
-      padding: .3em 0;
-      height: 1.6em;
-      border-radius: 4px;
-    }
-    .heroes li.selected:hover {
-      background-color: #BBD8DC !important;
-      color: white;
-    }
-    .heroes li:hover {
-      color: #607D8B;
-      background-color: #DDD;
-      left: .1em;
-    }
-    .heroes .text {
-      position: relative;
-      top: -3px;
-    }
-    .heroes .badge {
-      display: inline-block;
-      font-size: small;
-      color: white;
-      padding: 0.8em 0.7em 0 0.7em;
-      background-color: #607D8B;
-      line-height: 1em;
-      position: relative;
-      left: -1px;
-      top: -4px;
-      height: 1.8em;
-      margin-right: .8em;
-      border-radius: 4px 0 0 4px;
-    }
-  `],
-  //for another components to add into DOM
-  directives: [HeroDetailComponent],
-  // for services registration
-  // The providers array tells Angular to create a fresh instance of the HeroService when it creates a new
-  // AppComponent. The AppComponent can use that service to get heroes and so can every child component of its
-  // component tree.
-  providers: [HeroService]
+	selector: 'my-app',
+	template: `
+		<h1>{{title}}</h1>
+		<my-heroes></my-heroes>
+	`,
+	directives: [HeroesComponent],
+	providers: [
+		HeroService
+	]
 })
-export class AppComponent implements OnInit {
-  title = 'Tour of Heroes';
-  heroes: Hero[];
-  selectedHero: Hero;
-  
-  // to add new _heroService property and identifies it as a HeroService injection
-  constructor(private _heroService: HeroService) { }
-  
-  getHeroes() {
-    this._heroService.getHeroes().then(heroes => this.heroes = heroes);
-    // or we can use slow mode:)
-    // this._heroService.getHeroes().then(heroes => this.heroes = heroes);
-  }
 
-  // do something when initialize
-  ngOnInit() {
-    this.getHeroes();
-  }
-  onSelect(hero: Hero) { this.selectedHero = hero; }
+export class AppComponent {
+	title = 'Tour of Heroes';
 }
